+++ Sources/Navigator/Navigator.swift
@@ -47,29 +47,29 @@ public protocol Navigator {
 }
 
 public extension Navigator {
-    
+
     /// Adds default values for the parameters.
     @discardableResult
     func go(to locator: Locator, animated: Bool = false, completion: @escaping () -> Void = {}) -> Bool {
-        return go(to: locator, animated: animated, completion: completion)
+        go(to: locator, animated: animated, completion: completion)
     }
     
     /// Adds default values for the parameters.
     @discardableResult
     func go(to link: Link, animated: Bool = false, completion: @escaping () -> Void = {}) -> Bool {
-        return go(to: link, animated: animated, completion: completion)
+        go(to: link, animated: animated, completion: completion)
     }
     
     /// Adds default values for the parameters.
     @discardableResult
     func goForward(animated: Bool = false, completion: @escaping () -> Void = {}) -> Bool {
-        return goForward(animated: animated, completion: completion)
+        goForward(animated: animated, completion: completion)
     }
     
     /// Adds default values for the parameters.
     @discardableResult
     func goBackward(animated: Bool = false, completion: @escaping () -> Void = {}) -> Bool {
-        return goBackward(animated: animated, completion: completion)
+        goBackward(animated: animated, completion: completion)
     }
 
 }