+++ Sources/Navigator/VisualNavigator.swift
@@ -29,10 +29,18 @@ public protocol VisualNavigator: Navigator {
     @discardableResult
     func goRight(animated: Bool, completion: @escaping () -> Void) -> Bool
 
+    /// Returns the `Locator` to the first content element that begins on the current screen.
+    func firstVisibleElementLocator(completion: @escaping (Locator?) -> Void)
 }
 
 public extension VisualNavigator {
-    
+
+    func firstVisibleElementLocator(completion: @escaping (Locator?) -> ()) {
+        DispatchQueue.main.async {
+            completion(currentLocation)
+        }
+    }
+
     @discardableResult
     func goLeft(animated: Bool = false, completion: @escaping () -> Void = {}) -> Bool {
         switch readingProgression {
@@ -52,7 +60,6 @@ public extension VisualNavigator {
             return goBackward(animated: animated, completion: completion)
         }
     }
-    
 }
 
 
@@ -61,7 +68,6 @@ public protocol VisualNavigatorDelegate: NavigatorDelegate {
     /// Called when the user tapped the publication, and it didn't trigger any internal action.
     /// The point is relative to the navigator's view.
     func navigator(_ navigator: VisualNavigator, didTapAt point: CGPoint)
-    
 }
 
 public extension VisualNavigatorDelegate {
@@ -69,5 +75,4 @@ public extension VisualNavigatorDelegate {
     func navigator(_ navigator: VisualNavigator, didTapAt point: CGPoint) {
         // Optional
     }
-    
 }