+++ TestApp/Sources/Reader/Common/ReaderViewController.swift
@@ -1,13 +1,7 @@
 //
-//  ReaderViewController.swift
-//  r2-testapp-swift
-//
-//  Created by Mickaël Menu on 07.03.19.
-//
-//  Copyright 2019 European Digital Reading Lab. All rights reserved.
-//  Licensed to the Readium Foundation under one or more contributor license agreements.
-//  Use of this source code is governed by a BSD-style license which is detailed in the
-//  LICENSE file present in the project repository where this source code is maintained.
+//  Copyright 2022 Readium Foundation. All rights reserved.
+//  Use of this source code is governed by the BSD-style license
+//  available in the top-level LICENSE file of the project.
 //
 
 import Combine
@@ -33,10 +27,13 @@ class ReaderViewController: UIViewController, Loggable {
     private(set) var stackView: UIStackView!
     private lazy var positionLabel = UILabel()
     private var subscriptions = Set<AnyCancellable>()
-    
+
     private var searchViewModel: SearchViewModel?
     private var searchViewController: UIHostingController<SearchView>?
-    
+
+    private let ttsViewModel: TTSViewModel?
+    private let ttsControlsViewController: UIHostingController<TTSControls>?
+
     /// This regex matches any string with at least 2 consecutive letters (not limited to ASCII).
     /// It's used when evaluating whether to display the body of a noteref referrer as the note's title.
     /// I.e. a `*` or `1` would not be used as a title, but `on` or `好書` would.
@@ -55,9 +52,12 @@ class ReaderViewController: UIViewController, Loggable {
         self.books = books
         self.bookmarks = bookmarks
         self.highlights = highlights
-        
+
+        ttsViewModel = TTSViewModel(navigator: navigator, publication: publication)
+        ttsControlsViewController = ttsViewModel.map { UIHostingController(rootView: TTSControls(viewModel: $0)) }
+
         super.init(nibName: nil, bundle: nil)
-        
+
         addHighlightDecorationsObserverOnce()
         updateHighlightDecorations()
     
@@ -99,7 +99,7 @@ class ReaderViewController: UIViewController, Loggable {
         addChild(navigator)
         stackView.addArrangedSubview(navigator.view)
         navigator.didMove(toParent: self)
-        
+
         stackView.addArrangedSubview(accessibilityToolbar)
 
         positionLabel.translatesAutoresizingMaskIntoConstraints = false
@@ -110,6 +110,25 @@ class ReaderViewController: UIViewController, Loggable {
             positionLabel.centerXAnchor.constraint(equalTo: view.centerXAnchor),
             positionLabel.bottomAnchor.constraint(equalTo: navigator.view.bottomAnchor, constant: -20)
         ])
+
+        if let state = ttsViewModel?.$state, let controls = ttsControlsViewController {
+            controls.view.backgroundColor = .clear
+
+            addChild(controls)
+            controls.view.translatesAutoresizingMaskIntoConstraints = false
+            view.addSubview(controls.view)
+            NSLayoutConstraint.activate([
+                controls.view.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+                controls.view.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 20),
+            ])
+            controls.didMove(toParent: self)
+
+            state
+                .sink { state in
+                    controls.view.isHidden = !state.showControls
+                }
+                .store(in: &subscriptions)
+        }
     }
     
     override func willMove(toParent parent: UIViewController?) {
@@ -137,10 +156,13 @@ class ReaderViewController: UIViewController, Loggable {
         }
         // Bookmarks
         buttons.append(UIBarButtonItem(image: #imageLiteral(resourceName: "bookmark"), style: .plain, target: self, action: #selector(bookmarkCurrentPosition)))
-        
         // Search
         if publication._isSearchable {
             buttons.append(UIBarButtonItem(image: UIImage(systemName: "magnifyingglass"), style: .plain, target: self, action: #selector(showSearchUI)))
+        }
+        // Text to speech
+        if let ttsViewModel = ttsViewModel {
+            buttons.append(UIBarButtonItem(image: UIImage(systemName: "speaker.wave.2.fill"), style: .plain, target: ttsViewModel, action: #selector(TTSViewModel.start)))
         }
         
         return buttons
@@ -215,6 +237,7 @@ class ReaderViewController: UIViewController, Loggable {
     }
     
     // MARK: - Search
+
     @objc func showSearchUI() {
         if searchViewModel == nil {
             searchViewModel = SearchViewModel(publication: publication)
@@ -237,7 +260,7 @@ class ReaderViewController: UIViewController, Loggable {
         present(vc, animated: true, completion: nil)
         searchViewController = vc
     }
-    
+
     // MARK: - Highlights
     
     private func addHighlightDecorationsObserverOnce() {