+++ CHANGELOG.md
@@ -11,6 +11,11 @@ All notable changes to this project will be documented in this file. Take a look
 #### Shared
 
 * Support for the accessibility metadata in RWPM per [Schema.org Accessibility Properties for Discoverability Vocabulary](https://www.w3.org/2021/a11y-discov-vocab/latest/).
+* [Extract the raw content (text, images, etc.) of a publication](Documentation/Guides/Content.md).
+
+#### Navigator
+
+* [A brand new text-to-speech implementation](Documentation/Guides/TTS.md).
 
 ### Deprecated
 
@@ -25,6 +30,8 @@ All notable changes to this project will be documented in this file. Take a look
 
 * Fixed memory leaks in the EPUB and PDF navigators.
 * [#61](https://github.com/readium/swift-toolkit/issues/61) Fixed serving EPUB resources when the HREF contains an anchor or query parameters.
+* Performance issue with EPUB fixed-layout when spreads are enabled.
+* Disable scrolling in EPUB fixed-layout resources, in case the viewport is incorrectly set.
 
 #### Streamer